用户界面-商品收藏

+++ database/migrations/2019_10_31_073105_create_product_skus_table.php
@@ -14,12 +14,12 @@ class CreateProductSkusTable extends Migration
     public function up()
     {
         Schema::create('product_skus', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('title');
             $table->string('description');
             $table->decimal('price',10,2);
             $table->unsignedInteger('stock');
-            $table->unsignedBigInteger('product_id');
+            $table->unsignedInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
